server/requirements.txt
fastapi==0.115.4
uvicorn[standard]==0.30.6
python-multipart==0.0.12

# LlamaIndex split packages
llama-index-core
llama-index-llms-ollama
llama-index-embeddings-ollama
llama-index-readers-file
llama-index-vector-stores-weaviate
llama-index-llms-openai

weaviate-client>=3.26.7,<4.0.0
